Add a new annotation @MethodUnderTest
Defined in a new subpackage "generated" to emphasise that this is not to be used by the user but assigned by Diffblue Cover itself.
Using runtime retention - as I'm not sure is class would be enough for our purposes.
The value parameter has no default value, to make sure that the annotation is only ever used with a value.

Note: some of the other annotations are repeatable, this one should not be (only one annotation per test method should be used)